requirements.txt
Django==3.2
djangorestframework==3.13.1
drf-yasg==1.20.0
django-cors-headers==3.11.0
django-location-field==2.1.0